move getHostEventString to it's own file
